+++ PadeLovers/AppDelegate.swift
@@ -28,6 +28,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         LoginManager.shared.setup(channelID: "1657150279", universalLinkURL: nil)
 
         FirebaseApp.configure()
+
+        // UserDefaultsセットアップ
+        UserDefaultsUtil.shared.appLastLaunchDate = Date()
+        UserDefaultsUtil.shared.appLaunchCount += 1
         
         return true
     }
