+++ PadeLovers/Service/UserDefaultsUtil.swift
@@ -1,0 +1,64 @@
+//
+//  UserDefaultsManager.swift
+//  PadeLovers
+//
+//  Created by Yoshitaka Tanaka on 2022/06/11.
+//  Copyright © 2022 Yoshitaka. All rights reserved.
+//
+
+import Foundation
+
+final class UserDefaultsUtil {
+    static let shared = UserDefaultsUtil()
+    private var userDefaults = UserDefaults.standard
+
+    private init() { }
+
+    enum StringKey: String {
+        case padelId = "PadelID"
+        case newCourt
+        case appLastLaunchDate
+        case appLaunchCount
+    }
+}
+
+extension UserDefaultsUtil {
+    var padelId: String? {
+        get {
+            return UserDefaults.standard.string(forKey: StringKey.padelId.rawValue)
+        }
+        set {
+            guard let newValue = newValue else { return }
+            UserDefaults.standard.set(newValue, forKey: StringKey.padelId.rawValue)
+        }
+    }
+    var courtNames: [String]? {
+        get {
+            return UserDefaults.standard.value(forKey: StringKey.newCourt.rawValue) as? [String]
+        }
+        set {
+            guard let newValue = newValue else { return }
+            UserDefaults.standard.set(newValue, forKey: StringKey.newCourt.rawValue)
+        }
+    }
+
+    var appLastLaunchDate: Date {
+        get {
+            return Date(
+                timeIntervalSince1970: TimeInterval(userDefaults.integer(forKey: StringKey.appLastLaunchDate.rawValue))
+            )
+        }
+        set {
+            userDefaults.set(newValue.timeIntervalSince1970, forKey: StringKey.appLastLaunchDate.rawValue)
+        }
+    }
+
+    var appLaunchCount: Int {
+        get {
+            return userDefaults.integer(forKey: StringKey.appLaunchCount.rawValue)
+        }
+        set {
+            userDefaults.set(newValue, forKey: StringKey.appLastLaunchDate.rawValue)
+        }
+    }
+}
